+++ requirements.txt
@@ -1,0 +1,7 @@
+networkx==2.2
+setuptools==40.8.0
+tensorflow==1.13.1
+matplotlib==3.0.3
+numpy==1.16.2
+seaborn==0.9.0
+scikit_learn==0.20.3
